medical_institutions/requirements.txt
requests>=2.31.0
pandas>=2.2.0
beautifulsoup4>=4.12.3
pdfplumber>=0.11.0
fuzzywuzzy>=0.18.0
psycopg2-binary>=2.9.9
smart_open>=7.0.4
loguru>=0.7.2
python-Levenshtein>=0.25.0
lxml>=5.1.0
openpyxl>=3.1.2